Clipboard API: Implemented the HTML clipboard reader and writer
I added two subclasses, a clipboard html reader and a clipboard html
writer. When HTML is read and written it is "sanitized" which means
that all script objects are removed from the html. They are both
entirely synchronous because it is impossible to sanitize the html off
of the main thread.

Change-Id: I29b26f1c721102005bcbb5911549a9546aed3006

Bug: 931839
Change-Id: I29b26f1c721102005bcbb5911549a9546aed3006